Time to make a change
Need to make a hard descision and cut the losses of this very broken
cookbook. In order for the community at large to move on we need
to modernize this cookbook and make it a joy to use and work with
once again. I'm personally very sorry for my time away from this
cookbook, but feel the only way to get back is to break a few things
and modernize this cookbook to the Resource based cookbook model.

Want to help? Feel free to submit pull requests to help with this
effort.